rename: classCard -> classPreview
