+++ src/components/ClassPreview/ResponsiveClassPreview.tsx
